Remove ete test, add separate integration test class
