--- atlasdb-ete-tests/src/main/java/com/palantir/atlasdb/rows/GetRowKeysResource.java
@@ -1,46 +1,0 @@
-/*
- * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.atlasdb.rows;
-
-import java.util.List;
-
-import javax.ws.rs.Consumes;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.MediaType;
-
-import com.palantir.atlasdb.keyvalue.api.Cell;
-
-@Path("rows")
-public interface GetRowKeysResource {
-    @POST
-    @Path("reset-table")
-    void resetTable();
-
-    @POST
-    @Path("insert")
-    @Consumes(MediaType.APPLICATION_JSON)
-    void insertCell(Cell cell);
-
-    @POST
-    @Path("get")
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.APPLICATION_JSON)
-    List<byte[]> getRowKeys(int startRow, @QueryParam("maxRows") int maxRows);
-}
